requirements.txt
Flask==2.3.0
google-generativeai==0.8.3  
PyAudio==0.2.14
google-cloud-speech==2.29.0
google-cloud-texttospeech==2.23.0
python-dotenv==1.0.0
pydub==0.25.1
sounddevice==0.5.1
soundfile==0.12.1
google-auth==2.37.0
google-generativeai==0.8.3